# openapi.json
{"openapi":"3.1.0","info":{"title":"🚀 API Financeira Pro","description":"\n    API para consulta de ativos financeiros, criptomoedas e stocks americanas.\n\n    ## Endpoints Principais:\n    * **Search**: Busca global de ativos.\n    * **Quote**: Preço e tendência em tempo real.\n    * **History**: Dados para construção de gráficos.\n    * **News**: Notícias com resumo e fonte.\n    * **Dividends**: Histórico de proventos.\n    ","contact":{"name":"Seu Nome","url":"http://seu-link-ou-github.com/"},"version":"1.0.0"},"paths":{"/auth/login":{"post":{"tags":["Auth"],"summary":"Login For Access Token","operationId":"login_for_access_token_auth_login_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_for_access_token_auth_login_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/user/register":{"post":{"tags":["User"],"summary":"Create User","operationId":"create_user_user_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/user/delete/{user_id}":{"delete":{"tags":["User"],"summary":"Delete User","operationId":"delete_user_user_delete__user_id__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string","title":"User Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/user/update/{user_id}":{"put":{"tags":["User"],"summary":"Update User","operationId":"update_user_user_update__user_id__put","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string","title":"User Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/assets":{"get":{"tags":["Assets"],"summary":"List Assets","description":"Retorna a lista de ativos. Se 'search' for enviado, busca na base global do Yahoo.","operationId":"list_assets_assets_get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"search","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Search"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Asset"},"title":"Response List Assets Assets Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/assets/{ticker}":{"get":{"tags":["Assets"],"summary":"Get Asset","description":"Retorna os detalhes em tempo real de um ativo.","operationId":"get_asset_assets__ticker__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"ticker","in":"path","required":true,"schema":{"type":"string","title":"Ticker"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssetQuote"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/assets/{ticker}/history":{"get":{"tags":["Assets"],"summary":"Get Asset History","operationId":"get_asset_history_assets__ticker__history_get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"ticker","in":"path","required":true,"schema":{"type":"string","title":"Ticker"}},{"name":"period","in":"query","required":false,"schema":{"type":"string","default":"1mo","title":"Period"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/HistoryPoint"},"title":"Response Get Asset History Assets  Ticker  History Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/assets/{ticker}/financials":{"get":{"tags":["Assets"],"summary":"Get Asset Financials","operationId":"get_asset_financials_assets__ticker__financials_get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"ticker","in":"path","required":true,"schema":{"type":"string","title":"Ticker"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Financials"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/assets/{ticker}/dividends":{"get":{"tags":["Assets"],"summary":"Get Asset Dividends","operationId":"get_asset_dividends_assets__ticker__dividends_get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"ticker","in":"path","required":true,"schema":{"type":"string","title":"Ticker"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Dividend"},"title":"Response Get Asset Dividends Assets  Ticker  Dividends Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/assets/{ticker}/news":{"get":{"tags":["Assets"],"summary":"Get Asset News","operationId":"get_asset_news_assets__ticker__news_get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"ticker","in":"path","required":true,"schema":{"type":"string","title":"Ticker"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/NewsItem"},"title":"Response Get Asset News Assets  Ticker  News Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/profile/watchlist/add":{"post":{"tags":["Profile"],"summary":"Add To Watchlist","operationId":"add_to_watchlist_profile_watchlist_add_post","requestBody":{"content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Asset"},"type":"array","title":"Assets"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/profile/watchlist/remove":{"post":{"tags":["Profile"],"summary":"Remove From Watchlist","operationId":"remove_from_watchlist_profile_watchlist_remove_post","requestBody":{"content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/AssetRemove"},"type":"array","title":"Assets To Remove"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}}},"components":{"schemas":{"Asset":{"properties":{"ticker":{"type":"string","title":"Ticker"},"name":{"type":"string","title":"Name"},"icon_url":{"type":"string","title":"Icon Url"}},"type":"object","required":["ticker","name","icon_url"],"title":"Asset"},"AssetQuote":{"properties":{"ticker":{"type":"string","title":"Ticker"},"name":{"type":"string","title":"Name"},"icon_url":{"type":"string","title":"Icon Url"},"price_usd":{"type":"number","title":"Price Usd"},"direction":{"type":"string","title":"Direction"}},"type":"object","required":["ticker","name","icon_url","price_usd","direction"],"title":"AssetQuote"},"AssetRemove":{"properties":{"ticker":{"type":"string","title":"Ticker"}},"type":"object","required":["ticker"],"title":"AssetRemove"},"Body_login_for_access_token_auth_login_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"^password$"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","format":"password","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"format":"password","title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_for_access_token_auth_login_post"},"Dividend":{"properties":{"date":{"type":"string","title":"Date"},"amount":{"type":"number","title":"Amount"}},"type":"object","required":["date","amount"],"title":"Dividend"},"Financials":{"properties":{"market_cap":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Market Cap"},"pe_ratio":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Pe Ratio"},"dividend_yield":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Dividend Yield"},"target_price":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Target Price"},"recommendation":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Recommendation"}},"type":"object","required":["market_cap","pe_ratio","dividend_yield","target_price","recommendation"],"title":"Financials"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HistoryPoint":{"properties":{"date":{"type":"string","title":"Date"},"close":{"type":"number","title":"Close"}},"type":"object","required":["date","close"],"title":"HistoryPoint"},"NewsItem":{"properties":{"title":{"type":"string","title":"Title"},"link":{"type":"string","title":"Link"},"publisher":{"type":"string","title":"Publisher"},"summary":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Summary","default":""},"provider_publish_time":{"type":"string","title":"Provider Publish Time"}},"type":"object","required":["title","link","publisher","provider_publish_time"],"title":"NewsItem"},"UserCreate":{"properties":{"name":{"type":"string","title":"Name"},"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"},"birth_date":{"type":"string","format":"date","title":"Birth Date"},"investor_profile":{"type":"string","enum":["CONSERVATIVE","MODERATE","AGGRESSIVE"],"title":"Investor Profile"}},"type":"object","required":["name","email","password","birth_date","investor_profile"],"title":"UserCreate"},"UserResponse":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"name":{"type":"string","title":"Name"},"email":{"type":"string","format":"email","title":"Email"},"birth_date":{"type":"string","format":"date","title":"Birth Date"},"investor_profile":{"type":"string","title":"Investor Profile"},"is_active":{"type":"boolean","title":"Is Active"}},"type":"object","required":["id","name","email","birth_date","investor_profile","is_active"],"title":"UserResponse"},"UserUpdate":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"name":{"type":"string","title":"Name"},"email":{"type":"string","format":"email","title":"Email"},"investor_profile":{"type":"string","enum":["CONSERVATIVE","MODERATE","AGGRESSIVE"],"title":"Investor Profile"}},"type":"object","required":["id","name","email","investor_profile"],"title":"UserUpdate"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"auth/login"}}}}}}